refac: storage event globally set for user_templates & ketcher_user_settings
