+++ app/packs/src/apps/mydb/App.js
@@ -15,6 +15,10 @@ import UIStore from 'src/stores/alt/stores/UIStore';
 import UserActions from 'src/stores/alt/actions/UserActions';
 import Calendar from 'src/components/calendar/Calendar';
 import SampleTaskInbox from 'src/components/sampleTaskInbox/SampleTaskInbox';
+import UsersFetcher from 'src/fetchers/UsersFetcher';
+import ProfilesFetcher from '../../fetchers/ProfilesFetcher';
+
+const key = 'ketcher-tmpls';
 
 class App extends Component {
   constructor(_props) {
@@ -38,7 +42,7 @@ class App extends Component {
     UserActions.fetchOlsChmo();
     UserActions.fetchOlsBao();
     UserActions.fetchProfile();
-    UserActions.setUsertemplates()
+    UserActions.setUsertemplates();
     UserActions.fetchUserLabels();
     UserActions.fetchGenericEls();
     UserActions.fetchSegmentKlasses();
@@ -47,14 +51,109 @@ class App extends Component {
     UserActions.fetchEditors();
     UserActions.fetchKetcher2Options();
     UIActions.initialize.defer();
-    document.addEventListener('keydown', this.documentKeyDown);
-
     this.patchExternalLibraries();
+
+    document.addEventListener('keydown', this.documentKeyDown);
+    window.addEventListener('storage', this.handleStorageChange);
+
+    // user templates
+    this.removeLocalStorageEventListener();
+    this.storageListener();
+
   }
 
   componentWillUnmount() {
     UIStore.unlisten(this.handleUiStoreChange);
     document.removeEventListener('keydown', this.documentKeyDown);
+  }
+
+  removeLocalStorageEventListener() {
+    window.removeEventListener('storage', this.onChangeKetcherTemplates);
+  }
+
+  storageListener() {
+    window.addEventListener(
+      'storage',
+      // this.debounce(this.onChangeKetcherTemplates.bind(this), 300),
+      this.onEventListen.bind(this),
+      false
+    );
+  }
+
+  // helpers end
+  async onEventListen(event) {
+    let { newValue, oldValue } = event;
+    newValue = JSON.parse(newValue);
+    oldValue = JSON.parse(oldValue);
+    // const { deleteAllowed } = this.state;
+    if (event.key === key) { // matching key && deleteAllowed
+      if (newValue.length > oldValue.length) { // when a new template is added
+        console.log("new");
+        let newItem = newValue[newValue.length - 1];
+        this.createAddAttachmentidToNewUserTemplate(newValue, newItem);
+      } else if (newValue.length < oldValue.length) { // when a template is deleted
+        console.log("removed");
+        const listOfLocalid = newValue.map((item) => item.props.path);
+        this.removeUserTemplate(listOfLocalid, oldValue);
+      } else if (newValue.length == oldValue.length) { // when a template is update atom id, bond id
+        this.updateUserTemplateDetails(oldValue, newValue);
+      }
+    } else if (event.key === 'ketcher-opts') {
+      UsersFetcher.updateUserKetcher2Options(event.newValue);
+    }
+    else {
+      // this.setState({ deleteAllowed: true });
+    }
+  }
+
+  async createAddAttachmentidToNewUserTemplate(newValue, newItem, deleteIdx) {
+    const res = await ProfilesFetcher.uploadUserTemplates({
+      content: JSON.stringify(newItem),
+    }).catch(err => console.log("err in create"));
+    const attachment_id = res?.template_details?.filename;
+    newItem['props']['path'] = attachment_id;
+    newValue[newValue.length - 1] = newItem;
+    if (deleteIdx) newValue.splice(deleteIdx, 1);
+    // this.setState({ deleteAllowed: false });
+    this.removeLocalStorageEventListener();
+    localStorage.setItem(key, JSON.stringify(newValue));
+    // this.localStorageEventListener();
+  }
+
+
+  removeUserTemplate(listOfLocalid, oldValue) {
+    for (let i = 0; i < oldValue.length; i++) {
+      const localItem = oldValue[i];
+      const itemIndexShouldBeRemoved = listOfLocalid.indexOf(
+        localItem.props.path
+      );
+      if (itemIndexShouldBeRemoved === -1) {
+        ProfilesFetcher.deleteUserTemplate({
+          path: localItem?.props.path,
+        });
+        break;
+      }
+    }
+  }
+
+  async updateUserTemplateDetails(oldValue, newValue) {
+    const listOfLocalNames = newValue.map(
+      (item) => JSON.parse(item.struct).header.moleculeName
+    );
+    for (let i = 0; i < oldValue.length; i++) {
+      const localItem = JSON.parse(oldValue[i].struct);
+      const exists = listOfLocalNames.indexOf(localItem.header.moleculeName) !== -1;
+      if (!exists) {
+        console.log({ exists, name: localItem.header.moleculeName });
+        await ProfilesFetcher.deleteUserTemplate({
+          path: oldValue[i].props.path,
+        }).catch(() =>
+          console.log('ISSUE WITH DELETE', localItem?.props?.path)
+        );
+        this.createAddAttachmentidToNewUserTemplate(newValue, newValue[i], i);
+        break;
+      }
+    }
   }
 
   handleUiStoreChange(state) {
