--- db/migrate/20240617110459_add_user_templates_to_profiles.rb
@@ -1,5 +1,0 @@
-class AddUserTemplatesToProfiles < ActiveRecord::Migration[6.1]
-  def change
-    add_column :profiles, :user_templates, :string, array: true, default: []
-  end
-end
